requirements.txt
Django>=6.0
djangorestframework
djangorestframework-simplejwt
django-filter
drf-spectacular
psycopg2-binary
pytest-django
pytest
